requirements-macos.txt
numpy
keras-cv==0.4
tensorflow-datasets
tensorflow-macos==2.11.0
# tensorflow-metal
